Fix Black being unusable without editionless joker
The Black colour card was changed in the latest release to allow overriding other editions, but it was still checking for an editionless joker to determine whether it could be used at all. Now it can be used as long as you have at least one non-Negative joker.
Also, its effect will now target only non-Negative jokers; it won't be wasted on ones that were already Negative.